update: tax have displayTax

+++ dtos/receipt_dto.go
@@ -9,7 +9,7 @@ type ReceiptItem struct {
 type ReceiptResponse struct {
 	StoreName string        `json:"storeName" example:"Nigger Store"`
 	BillDate  string        `json:"billDate" example:"2025-10-02"`
-	Tax       float32       `json:"tax" example:"9.10"`
-	Service   float32       `json:"service"`
+	Tax       float64       `json:"tax" example:"9.10"`
+	Service   float64       `json:"service"`
 	Items     []ReceiptItem `json:"items"`
 }
